+++ src/pulumi/core-shared/__main__.py
@@ -1,0 +1,6 @@
+# All sub-components to load. Please note, this is not an execution order.
+# Pulumi will load all files first and then will build the dependency graph.
+import management
+import network
+import security
+import automation